[package] change sigscheme lib installation directory
This patch installs the lib files for sigscheme (.scm files) in
/usr/share/sigscheme/lib instead of /usr/lib/sigscheme/lib.

If installed in another location, sigscheme can't find them and
refuses to start.

Another way to tackle the problem would be to make sigscheme look
for the .scm files in /usr/lib/sigscheme (the trailinig /lib seems
more than a bit superfluous), but that would require patching
configure.ac, which is probably more trouble than it's worth.

Signed-off-by: Andrea Bolognani <eof@kiyuko.org>